Fix bugs in request and report handling.
1. Correctly detect source IP from request.
2. Remodularize requests into their own file (hoping not to recompile them unless they change).
3. Add missing fields in report headers.
4. Make the nginx configuration pass the client-supplied Host header so we can tell whether the request is meant for Adobe or not.
