+++ rsrc/install/ubuntu/adlu-proxy.conf
@@ -13,6 +13,7 @@ server {
             proxy_pass http://localhost:8080;
             proxy_redirect default;
 
+            proxy_set_header Host $http_host;
             proxy_set_header X-Real-IP $remote_addr;
             proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
             proxy_set_header X-Forwarded-Proto $scheme;
